Extract clockmock from clock
